# public/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Pizza Data System API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
        "schemas": {
            "ErrorResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "error": {
                        "type": "string"
                    },
                    "file": {
                        "type": "string"
                    },
                    "line": {
                        "type": "integer"
                    }
                },
                "required": ["success", "error"]
            },
...
            "CsvFileInfo": {
                "type": "object",
                "properties": {
                    "name": { "type": "string" },
                    "size": { "type": "integer", "nullable": true },
                    "size_mb": { "type": "number", "nullable": true }
                },
                "required": ["name"]
            },
            "InspectZipResponse": {
                "type": "object",
                "properties": {
                    "success": { "type": "boolean" },
                    "temp_id": { "type": "string" },
                    "csv_files": {
                        "type": "array",
                        "items": { "$ref": "#/components/schemas/CsvFileInfo" }
                    }
                },
                "required": ["success", "temp_id", "csv_files"]
            },
...
            "ReaggregateResponse": {
                "type": "object",
                "properties": {
                    "success": { "type": "boolean" },
                    "aggregation_id": { "type": "string" }
                },
                "required": ["success", "aggregation_id"]
            },
            "AggregationProgress": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string",
                        "enum": ["queued", "processing", "completed", "failed"]
                    },
                    "processed": { "type": "integer" },
                    "total": { "type": "integer" },
                    "started_at": { "type": "string", "format": "date-time" }
                },
                "required": ["status", "processed", "total", "started_at"]
            },
            "AggregationProgressResponse": {
                "type": "object",
                "properties": {
                    "success": { "type": "boolean" },
                    "progress": {
                        "$ref": "#/components/schemas/AggregationProgress"
                    }
                },
                "required": ["success", "progress"]
            },
            "GenericErrorResponse": {
                "type": "object",
                "properties": {
                    "success": { "type": "boolean", "example": false },
                    "message": { "type": "string" }
                },
                "required": ["success", "message"]
            }
        }
    },
    "security": [
        {
            "BearerAuth": []
        }
    ],
    "paths": {
        "/export/csv": {
            "get": {
                "summary": "Export data as CSV",
                "description": "Downloads a CSV file for a specific model or a ZIP containing CSVs for all models.",
                "operationId": "exportCsv",
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "start",
                        "in": "query",
                        "schema": { "type": "string", "format": "date" },
                        "required": false
                    },
                    {
                        "name": "end",
                        "in": "query",
                        "schema": { "type": "string", "format": "date" },
                        "required": false
                    },
                    {
                        "name": "store",
                        "in": "query",
                        "schema": { "type": "string" },
                        "required": false
                    },
                    {
                        "name": "model",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "detail_orders",
                                "order_line",
                                "summary_sales",
                                "summary_items",
                                "summary_transactions",
                                "waste",
                                "cash_management",
                                "financial_views",
                                "alta_inventory_waste",
                                "alta_inventory_ingredient_usage",
                                "alta_inventory_ingredient_orders",
                                "alta_inventory_cogs",
                                "yearly_store_summary",
                                "yearly_item_summary",
                                "weekly_store_summary",
                                "weekly_item_summary",
                                "quarterly_store_summary",
                                "quarterly_item_summary",
                                "monthly_store_summary",
                                "monthly_item_summary",
                                "daily_store_summary",
                                "daily_item_summary",
                                "hourly_store_summary",
                                "hourly_item_summary",
                                "all"
                            ]
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CSV or ZIP file download",
                        "content": {
                            "text/csv": {},
                            "application/zip": {}
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": true
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error during export",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/manual-import/inspect-zip": {
            "post": {
                "summary": "Inspect ZIP file and list CSV files",
                "operationId": "inspectZip",
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary",
                                        "description": "ZIP file containing CSV files (max 1GB)"
                                    }
                                },
                                "required": ["file"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ZIP inspected successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InspectZipResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/manual-import/reaggregate": {
            "post": {
                "summary": "Trigger re-aggregation",
                "description": "Re-aggregate data for a specific date range and aggregation type",
                "operationId": "reaggregate",
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "start_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "hourly",
                                            "daily",
                                            "weekly",
                                            "monthly",
                                            "quarterly",
                                            "yearly",
                                            "all"
                                        ]
                                    }
                                },
                                "required": ["start_date", "end_date", "type"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Re-aggregation initiated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReaggregateResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/manual-import/aggregation-progress/{aggregationId}": {
            "get": {
                "summary": "Get aggregation progress",
                "operationId": "aggregationProgress",
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "aggregationId",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Aggregation progress",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AggregationProgressResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aggregation ID not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}